数据组件调整

+++ c.main/c.dao.admin/src/main/java/com/lt/shop/dao/admin/entity/def/UserAddr.java
@@ -3,15 +3,15 @@ package com.lt.shop.dao.admin.entity.def;
 import java.util.Date;
 
 public class UserAddr {
-    private Integer id;
+    private Long id;
 
-    private Integer userId;
+    private Long userId;
 
-    private Integer provinceId;
+    private Long provinceId;
 
-    private Integer cityId;
+    private Long cityId;
 
-    private Integer districtId;
+    private Long districtId;
 
     private String address;
 
@@ -23,43 +23,43 @@ public class UserAddr {
 
     private Integer defaultUsed;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 
-    public Integer getProvinceId() {
+    public Long getProvinceId() {
         return provinceId;
     }
 
-    public void setProvinceId(Integer provinceId) {
+    public void setProvinceId(Long provinceId) {
         this.provinceId = provinceId;
     }
 
-    public Integer getCityId() {
+    public Long getCityId() {
         return cityId;
     }
 
-    public void setCityId(Integer cityId) {
+    public void setCityId(Long cityId) {
         this.cityId = cityId;
     }
 
-    public Integer getDistrictId() {
+    public Long getDistrictId() {
         return districtId;
     }
 
-    public void setDistrictId(Integer districtId) {
+    public void setDistrictId(Long districtId) {
         this.districtId = districtId;
     }
