+++ c.main/c.dao.admin/src/main/java/com/lt/shop/dao/site/impl/SiteOrderLogMapper.java
@@ -3,13 +3,13 @@ package com.lt.shop.dao.site.impl;
 import com.lt.shop.dao.admin.entity.def.OrderLog;
 
 public interface SiteOrderLogMapper {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(OrderLog record);
 
     int insertSelective(OrderLog record);
 
-    OrderLog selectByPrimaryKey(Integer id);
+    OrderLog selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(OrderLog record);
